feat: rename game to 'Lexi War' and update descriptions; refactor lobby creation to use CreateLobbyForm
